LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/events/event-handler-spec-example-expected.txt


FAIL Event handler listeners should be registered when they are first set to an object value (object "[object Object]"). assert_equals: expected 2 but got 1
FAIL Event handler listeners should be registered when they are first set to an object value (function "function () {}"). assert_equals: expected 2 but got 1
FAIL Event handler listeners should be registered when they are first set to an object value (object "42"). assert_equals: expected 2 but got 1
FAIL Event handler listeners should be registered when they are first set to an object value (object ""). assert_equals: expected 2 but got 1
FAIL Event handler listeners should be registered when they are first set to an object value (42). assert_equals: expected 3 but got 2
FAIL Event handler listeners should be registered when they are first set to an object value (null). assert_equals: expected 3 but got 2
FAIL Event handler listeners should be registered when they are first set to an object value (undefined). assert_equals: expected 3 but got 2
FAIL Event handler listeners should be registered when they are first set to an object value (""). assert_equals: expected 3 but got 2
FAIL Event handler listeners should be registered when they are first set to an object value. assert_equals: expected 3 but got 2
